# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Warehouse Inventory Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scan-in:
    get:
      operationId: listScanIn
      tags: [scan-in]
      summary: List all scan-in sessions
      parameters:
        - name: materialId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of scan-in sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScanIn"
    post:
      operationId: createScanIn
      tags: [scan-in]
      summary: Create a scan-in session (start scanning)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanInInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanIn"
  /scan-in/{id}:
    get:
      operationId: getScanIn
      tags: [scan-in]
      summary: Get scan-in session by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: ScanIn session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanIn"
        "404":
          description: Not found
    patch:
      operationId: updateScanIn
      tags: [scan-in]
      summary: Update scan-in session (e.g., finalize)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanInUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanIn"
    delete:
      operationId: deleteScanIn
      tags: [scan-in]
      summary: Delete a scan-in session (master only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /scan-in/{id}/items:
    post:
      operationId: addScanInItem
      tags: [scan-in]
      summary: Add a serial number to a scan-in session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanItemInput"
      responses:
        "201":
          description: Item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanItem"
        "409":
          description: Duplicate serial number
  /scan-in/{id}/items/{itemId}:
    delete:
      operationId: deleteScanInItem
      tags: [scan-in]
      summary: Remove an item from a scan-in session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /scan-out:
    get:
      operationId: listScanOut
      tags: [scan-out]
      summary: List all scan-out sessions
      parameters:
        - name: materialId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of scan-out sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScanOut"
    post:
      operationId: createScanOut
      tags: [scan-out]
      summary: Create a scan-out session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanOutInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanOut"
  /scan-out/{id}:
    get:
      operationId: getScanOut
      tags: [scan-out]
      summary: Get scan-out session by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: ScanOut session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanOut"
        "404":
          description: Not found
    delete:
      operationId: deleteScanOut
      tags: [scan-out]
      summary: Delete a scan-out session (master only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /scan-out/{id}/items:
    post:
      operationId: addScanOutItem
      tags: [scan-out]
      summary: Process a QR code scan for material-out
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanOutItemInput"
      responses:
        "201":
          description: Item processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanItem"
        "404":
          description: Serial number not found in stock
        "409":
          description: Already scanned out
...
components:
  schemas:
...
    ScanIn:
      type: object
      required: [id, materialId, materialName, boxLabel, status, userId, userName, createdAt, items]
      properties:
        id:
          type: integer
        materialId:
          type: integer
        materialName:
          type: string
        boxLabel:
          type: string
        status:
          type: string
          enum: [scanning, completed]
        userId:
          type: integer
        userName:
          type: string
        qrCodeData:
          type: ["string", "null"]
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/ScanItem"
    ScanInInput:
      type: object
      required: [materialId, userId]
      properties:
        materialId:
          type: integer
        userId:
          type: integer
    ScanInUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [scanning, completed]
        qrCodeData:
          type: string
    ScanOut:
      type: object
      required: [id, userId, userName, createdAt, items]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/ScanItem"
    ScanOutInput:
      type: object
      required: [userId]
      properties:
        userId:
          type: integer
    ScanOutItemInput:
      type: object
      required: [qrData]
      properties:
        qrData:
          type: string
    ScanItem:
      type: object
      required: [id, serialNumber, createdAt]
      properties:
        id:
          type: integer
        serialNumber:
          type: string
        scanInId:
          type: ["integer", "null"]
        scanOutId:
          type: ["integer", "null"]
        createdAt:
          type: string
    ScanItemInput:
      type: object
      required: [serialNumber]
      properties:
        serialNumber:
          type: string
          minLength: 28
          maxLength: 28
...